Update project to .NET 8.0 and refactor code to C# 10 syntax
This commit updates the target framework of the project to .NET 8.0 and the language version to C# 12 in all .csproj files. It also refactors the code to make use of C# 10's file-scoped namespaces and semicolon syntax for more concise code. A global.json file was also added to specify the SDK version.